Mashallah

+++ Frontend/src/components/workout.jsx
@@ -1,12 +1,23 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 function Workout({workout}){
 
     const {dispatch} = useWorkoutsContext()
+    const {user} = useAuthContext()
+
     const handleDelete = async() => {
+
+        if(!user){
+            return
+        }
+        
         const respose = await fetch("/api/workouts/" + workout._id ,{
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization' : `Bearer: ${user.token}`
+            }
         })
         const json = await respose.json()
         if (respose.ok){
